requirements.txt
fastapi==0.95.1
gdown==4.7.1
numpy==1.23.5
Pillow==9.5.0
pydantic==1.10.7
tensorflow==2.12.0
uvicorn==0.21.1
python-multipart==0.0.6
google-cloud-storage==2.9.0
mysql-connector==2.2.9